doc_assistant_django_v4/requirements.txt
Django>=5.2,<6.0
djangorestframework
djangorestframework-simplejwt
django-ratelimit
python-dotenv
openai
sentence-transformers
faiss-cpu
PyPDF2
tqdm
gunicorn
whitenoise
celery[redis]
redis
python-multipart
pytest
pytest-django
sentry-sdk

django-celery-beat
pinecone-client
